requirements.txt
chainlit==1.0.401
langchain==0.1.12
langchain_community==0.0.28
PyPDF2==3.0.1
python-dotenv==1.0.1
python_docx==1.1.0

openai
chromadb
tiktoken
pandas
tabulate
pypdf
sentence-transformers